primitive type migration from mod.rs to primitives.rs
